+++ src/sv_shell/run_whamg.sh
@@ -6,6 +6,11 @@
 
 set -Exeuo pipefail
 
+
+if [ -z "${SV_SHELL_CLEAN_UP_WORKING_DIR:-}" ]; then
+  SV_SHELL_CLEAN_UP_WORKING_DIR=true
+fi
+
 sample_id=$1
 cram_file=$2
 cram_index=$3
@@ -83,7 +88,9 @@ mv "${sample_id}.wham.vcf.gz" "${vcf_filename}"
 index_filename="${output_dir}/${sample_id}.wham.vcf.gz.tbi"
 mv "${sample_id}.wham.vcf.gz.tbi" "${index_filename}"
 
-rm -rf "${working_dir}"
+if [ "${SV_SHELL_CLEAN_UP_WORKING_DIR}" == "true" ]; then
+  rm -rf "${working_dir}"
+fi
 
 jq -n \
   --arg vcf "${vcf_filename}" \
